extract method and data to common.js
