+++ app/javascript/gobierto_data/webapp/components/sets/VisualizationsItem.vue
@@ -70,7 +70,7 @@ import { Loading } from "lib/vue/components";
 import Visualizations from "./../commons/Visualizations.vue";
 import SavingDialog from "./../commons/SavingDialog.vue";
 import { getUserId } from "./../../../lib/helpers";
-import * as htmlToImage from 'html-to-image';
+import { convertVizToImgMixin } from "./../../../lib/commons.js";
 
 export default {
   name: "VisualizationsItem",
@@ -79,6 +79,7 @@ export default {
     SavingDialog,
     Loading
   },
+  mixins: [convertVizToImgMixin],
   props: {
     datasetId: {
       type: Number,
@@ -270,22 +271,7 @@ export default {
       }
     },
     //TODO: Extract to common.js
-    convertVizToImg(opts) {
-      let node = document.querySelector('.gobierto-data-visualization--aspect-ratio-16-9');
-      const perspectiveChart = document.querySelector("perspective-viewer").shadowRoot
-      const perspectiveSidePanel = perspectiveChart.getElementById("side_panel")
-      const perspectiveTopPanel = perspectiveChart.getElementById("top_panel")
-      perspectiveSidePanel.style.display = "none"
-      perspectiveTopPanel.style.display = "none"
-      htmlToImage.toPng(node)
-        .then(function (dataUrl) {
-          this.imageApi = dataUrl
-          this.onSaveEventHandler(opts)
-        }.bind(this))
-        .catch(function (error) {
-          console.error('oops, something went wrong!', error);
-        });
-    },
+
     updateVizName(value) {
       const {
         name: vizName
