+++ app/javascript/gobierto_data/webapp/components/sets/data/SQLEditorResults.vue
@@ -95,7 +95,7 @@ import DownloadButton from "./../../commons/DownloadButton.vue";
 import DownloadLink from "./../../commons/DownloadLink.vue";
 import SavingDialog from "./../../commons/SavingDialog.vue";
 import Visualizations from "./../../commons/Visualizations.vue";
-import * as htmlToImage from 'html-to-image';
+import { convertVizToImgMixin } from "./../../../../lib/commons.js";
 
 export default {
   name: "SQLEditorResults",
@@ -106,6 +106,7 @@ export default {
     SavingDialog,
     DownloadLink
   },
+  mixins: [convertVizToImgMixin],
   props: {
     arrayFormats: {
       type: Object,
@@ -182,8 +183,7 @@ export default {
       removeLabelBtn: false,
       perspectiveChanged: false,
       config: null,
-      configMapZoom: { ...this.configMap, zoom: true },
-      imageApi: null
+      configMapZoom: { ...this.configMap, zoom: true }
     };
   },
   computed: {
@@ -216,26 +216,6 @@ export default {
       const config = this.$refs.viewer.getConfig()
       config.base64 = this.imageApi
       this.$root.$emit("storeCurrentVisualization", config, opts);
-    },
-    //TODO: Extract to common.js
-    convertVizToImg(opts) {
-      let node = document.querySelector('.gobierto-data-visualization--aspect-ratio-16-9');
-      const perspectiveChart = document.querySelector("perspective-viewer").shadowRoot
-      //Hide the top filters and the left sidebar
-      const perspectiveSidePanel = perspectiveChart.getElementById("side_panel")
-      const perspectiveTopPanel = perspectiveChart.getElementById("top_panel")
-      perspectiveSidePanel.style.display = "none"
-      perspectiveTopPanel.style.display = "none"
-      htmlToImage.toPng(node)
-        .then(function (dataUrl) {
-          this.imageApi = dataUrl
-          this.onSaveEventHandler(opts)
-          perspectiveSidePanel.style.display = "flex"
-          perspectiveTopPanel.style.display = "flex"
-        }.bind(this))
-        .catch(function (error) {
-          console.error('oops, something went wrong!', error);
-        });
     },
     updateVizNameHandler(value) {
       const {
